Remove hard-coded initial guess and lower bound for temperature.
The previous model change only removed the lower bound at the input stage, but the data process still has a hard-coded temperature bound and initial guess. This was initially needed when the model function was stiffer and would fail to fit. Removing this bound improves the worst-case fit of surface temperature even further.