Introduce stage to associate benchmarks with runs, between the runs and pipeline stages.
Modify the directory model to include a benchmark data source and an intermediate data store where runs and benchmarks will be combined. Introduce a new data processing stage that trivially loads experimental runs and saves them in the new intermediate location. Change the source of the main pipeline to the new intermediate location and reproduce the data process end-to-end. This produces all outputs exactly the same as before. However, now the trivial benchmark association stage may be modified to add new functionality to the data process.