Generate benchmark statistics useful for evaluating sensor performance
Remove the possibility of there being a sixth thermocouple feature and specify copper thermocouples in fixed parameters. This simplifies things and allows copper temperatures to be retrieved anywhere in the data process.

The increased expressivity of the data process facilitates parsing of benchmarks. The 90% rise time of each thermocouple is found, as well as the temperature of each thermocouple at the 90% rise time for the base thermocouple. Both give an indication of how well the thermocouples are installed.