Refine benchmark statistic
Base the rise time on the difference between the beginning and end temperatures of the base thermocouple, rather than the minimum and maximum temperatures. This reduces the influence of varying overshoot on the benchmark. It is apparent that the rise time for each thermocouple is not as relevant as the actual thermocouple temperature at the 90% rise time of the base thermocouple. Make this the only benchmark statistic.