Test notebooks. Test other parts of pipeline.
## Test notebooks

Facilitate running notebooks in a context where patched parameters apply. Initial attempts to do this with `testbook` led to issues. Old-style `unittest.mock.patch` patching is used to patch notebook values, and there are odd limitations with notebook interactions.

The alternative, `pytest-notebook` doesn't allow patching at all, but does handle coverage statistics. That would be nice to have, but my notebooks rely on patched content from other imports, so this doesn't work, either.

Finally, it was decided to just convert the notebook to a Python script and test that in the shared patch context set up in tests. This works, but it took awhile to settle on a vendor for notebook conversion. My choice to use `nbqa` internals proved time-consuming, and I could probably refactor the `get_nb_content` test helper function to use `nbconvert` library API instead.

Once notebooks were testable, special behavior enabled by auto-imports like `disp` and notebook history like `_i` were causing tests to fail. This magic behavior in notebooks is nice, but should not be used if it is desired to test the notebooks. At least `disp` can be explicitly imported to solve that problem. The tougher problem was replacing `_i` usage with tests that check `vars(syms)` of module `syms` for correct order of symbol assignment in groups. All of these symbolic variables can then be imported to the notebook and used as usual.

Finally, an actual bug in plotting was discovered, which only cropped up with small datasets like the test dataset. If an empty dataset is returned from a function, everything would fall apart.

## Test other parts of pipeline

Most of the pipeline is covered now. Programmatically parametrized stage tests makes this trivial as long as the pipeline stages are in the same package directory. The `stages` package will need to be flattened, and `common` modules refactored into the nearest `__init__.py`.

## Separate out dependencies that must be tested manually

This is no longer necessary, as most of the data process is now testable.